♻️ Move versioning into models
