+++ lamindb/models/transform.py
@@ -16,9 +16,9 @@ from lamindb.base.fields import (
 from lamindb.base.types import TransformType
 from lamindb.base.users import current_user_id
 
-from ..core.versioning import message_update_key_in_version_family, process_revises
 from ..errors import InconsistentKey
 from ..models._settings import settings
+from ..models._versioning import message_update_key_in_version_family, process_revises
 from .base import IsVersioned
 from .record import Record, init_self_from_db, update_attributes
 from .run import Run, User, delete_run_artifacts
